// graphql/annict/generated/graphql.ts
export type InputMaybe<T> = T | null | undefined;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: unknown; output: unknown; }
};
// ...
export type Media =
  | 'MOVIE'
  | 'OTHER'
  | 'OVA'
  | 'TV'
  | 'WEB';
// ...
export type SeasonName =
  | 'AUTUMN'
  | 'SPRING'
  | 'SUMMER'
  | 'WINTER';
// ...
export type ListWorksQueryVariables = Exact<{
  after?: InputMaybe<Scalars['String']['input']>;
}>;


export type ListWorksQuery = { searchWorks?: { nodes?: Array<{ annictId: number, malAnimeId?: string | null, syobocalTid?: number | null, title: string, media: Media, seasonYear?: number | null, seasonName?: SeasonName | null } | null> | null, pageInfo: { hasNextPage: boolean, endCursor?: string | null } } | null };


export const ListWorksDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"listWorks"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"after"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"searchWorks"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"after"},"value":{"kind":"Variable","name":{"kind":"Name","value":"after"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"annictId"}},{"kind":"Field","name":{"kind":"Name","value":"malAnimeId"}},{"kind":"Field","name":{"kind":"Name","value":"syobocalTid"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"media"}},{"kind":"Field","name":{"kind":"Name","value":"seasonYear"}},{"kind":"Field","name":{"kind":"Name","value":"seasonName"}}]}},{"kind":"Field","name":{"kind":"Name","value":"pageInfo"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"hasNextPage"}},{"kind":"Field","name":{"kind":"Name","value":"endCursor"}}]}}]}}]}}]} as unknown as DocumentNode<ListWorksQuery, ListWorksQueryVariables>;